Split out business logic and boilerplate
